拆分方法
